pass request options as struct
